# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface AuthUser {
  id: string;
  username?: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string;
  role: AuthUserRole;
}

export interface GetCurrentAuthUserResponse {
  isAuthenticated: boolean;
  user?: AuthUser;
}

export type AdminUserRole = (typeof AdminUserRole)[keyof typeof AdminUserRole];

export const AdminUserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface AdminUser {
  id: string;
  username?: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string;
  role: AdminUserRole;
  createdAt?: string;
}

export type UpdateUserRoleBodyRole =
  (typeof UpdateUserRoleBodyRole)[keyof typeof UpdateUserRoleBodyRole];

export const UpdateUserRoleBodyRole = {
  user: "user",
  admin: "admin",
} as const;

export interface UpdateUserRoleBody {
  role: UpdateUserRoleBodyRole;
}

export type CharacterInventarioItem = { [key: string]: unknown };

export interface RpgClass {
  id: string;
  nome: string;
  descricao?: string;
  pvInicial?: number;
  pvPorNivel?: number;
  peInicial?: number;
  pePorNivel?: number;
  sanInicial?: number;
  sanPorNivel?: number;
  periciasTreindasBase?: number;
  createdAt?: string;
}

export interface Origin {
  id: string;
  nome: string;
  descricao?: string;
  periciasConcedidas?: string[];
  poderConcedido?: string;
  createdAt?: string;
}

export interface Character {
  id: string;
  userId: string;
  nome: string;
  classeId?: string;
  origemId?: string;
  nivel: number;
  nex: number;
  patente?: string;
  pvAtual?: number;
  pvMaximo?: number;
  peAtual?: number;
  peMaximo?: number;
  sanAtual?: number;
  sanMaximo?: number;
  forca?: number;
  agilidade?: number;
  intelecto?: number;
  vigor?: number;
  presenca?: number;
  defesa?: number;
  historia?: string;
  pericias?: string[];
  rituals?: string[];
  inventario?: CharacterInventarioItem[];
  classe?: RpgClass;
  origem?: Origin;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreateCharacterBody {
  nome: string;
  classeId?: string;
  origemId?: string;
  nivel?: number;
  nex?: number;
  patente?: string;
  forca?: number;
  agilidade?: number;
  intelecto?: number;
  vigor?: number;
  presenca?: number;
  historia?: string;
  pericias?: string[];
}

export type UpdateCharacterBodyInventarioItem = { [key: string]: unknown };

export interface UpdateCharacterBody {
  nome?: string;
  classeId?: string;
  origemId?: string;
  nivel?: number;
  nex?: number;
  patente?: string;
  pvAtual?: number;
  peAtual?: number;
  sanAtual?: number;
  forca?: number;
  agilidade?: number;
  intelecto?: number;
  vigor?: number;
  presenca?: number;
  defesa?: number;
  historia?: string;
  pericias?: string[];
  rituals?: string[];
  inventario?: UpdateCharacterBodyInventarioItem[];
}

export interface CreateClassBody {
  nome: string;
  descricao?: string;
  pvInicial: number;
  pvPorNivel: number;
  peInicial: number;
  pePorNivel: number;
  sanInicial: number;
  sanPorNivel: number;
  periciasTreindasBase?: number;
}

export interface CreateOriginBody {
  nome: string;
  descricao?: string;
  periciasConcedidas?: string[];
  poderConcedido?: string;
}

export type PericiaAtributoBase =
  (typeof PericiaAtributoBase)[keyof typeof PericiaAtributoBase];

export const PericiaAtributoBase = {
  FOR: "FOR",
  AGI: "AGI",
  INT: "INT",
  VIG: "VIG",
  PRE: "PRE",
} as const;

export interface Pericia {
  id: string;
  nome: string;
  atributoBase: PericiaAtributoBase;
  somenteTrainada?: boolean;
  createdAt?: string;
}

export type CreatePericiaBodyAtributoBase =
  (typeof CreatePericiaBodyAtributoBase)[keyof typeof CreatePericiaBodyAtributoBase];

export const CreatePericiaBodyAtributoBase = {
  FOR: "FOR",
  AGI: "AGI",
  INT: "INT",
  VIG: "VIG",
  PRE: "PRE",
} as const;

export interface CreatePericiaBody {
  nome: string;
  atributoBase: CreatePericiaBodyAtributoBase;
  somenteTrainada?: boolean;
}

export type RitualElemento =
  (typeof RitualElemento)[keyof typeof RitualElemento];

export const RitualElemento = {
  Sangue: "Sangue",
  Morte: "Morte",
  Conhecimento: "Conhecimento",
  Energia: "Energia",
  Medo: "Medo",
} as const;

export interface Ritual {
  id: string;
  nome: string;
  elemento: RitualElemento;
  circulo: number;
  execucao?: string;
  alcance?: string;
  alvo?: string;
  duracao?: string;
  resistencia?: string;
  custoPe?: number;
  descricao?: string;
  createdAt?: string;
}

export type CreateRitualBodyElemento =
  (typeof CreateRitualBodyElemento)[keyof typeof CreateRitualBodyElemento];

export const CreateRitualBodyElemento = {
  Sangue: "Sangue",
  Morte: "Morte",
  Conhecimento: "Conhecimento",
  Energia: "Energia",
  Medo: "Medo",
} as const;

export interface CreateRitualBody {
  nome: string;
  elemento: CreateRitualBodyElemento;
  circulo: number;
  execucao?: string;
  alcance?: string;
  alvo?: string;
  duracao?: string;
  resistencia?: string;
  custoPe?: number;
  descricao?: string;
}

export type ItemTipo = (typeof ItemTipo)[keyof typeof ItemTipo];

export const ItemTipo = {
  Arma: "Arma",
  Protecao: "Protecao",
  Utilitario: "Utilitario",
  Amaldicoado: "Amaldicoado",
  Outro: "Outro",
} as const;

export type ItemRequisitos = { [key: string]: unknown };

export interface Item {
  id: string;
  nome: string;
  tipo: ItemTipo;
  descricao?: string;
  espacos?: number;
  categoria?: string;
  peso?: number;
  preco?: number;
  requisitos?: ItemRequisitos;
  createdAt?: string;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export type CreateItemBodyTipo =
  (typeof CreateItemBodyTipo)[keyof typeof CreateItemBodyTipo];

export const CreateItemBodyTipo = {
  Arma: "Arma",
  Protecao: "Protecao",
  Utilitario: "Utilitario",
  Amaldicoado: "Amaldicoado",
  Outro: "Outro",
} as const;

export type CreateItemBodyRequisitos = { [key: string]: unknown };

export interface CreateItemBody {
  nome: string;
  tipo: CreateItemBodyTipo;
  descricao?: string;
  espacos?: number;
  categoria?: string;
  peso?: number;
  preco?: number;
  requisitos?: CreateItemBodyRequisitos;
}