# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ordo Realitas RPG API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List all characters for authenticated user
      responses:
        "200":
          description: List of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
        "401":
          description: Unauthorized
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a new character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterBody"
      responses:
        "201":
          description: Character created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "401":
          description: Unauthorized
  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get character by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Character details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found
    put:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharacterBody"
      responses:
        "200":
          description: Character updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Exchange mobile authorization code for session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody"
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse"
...
components:
  schemas:
...
    Character:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        nome:
          type: string
        classeId:
          type: string
        origemId:
          type: string
        nivel:
          type: integer
        nex:
          type: integer
        patente:
          type: string
        pvAtual:
          type: integer
        pvMaximo:
          type: integer
        peAtual:
          type: integer
        peMaximo:
          type: integer
        sanAtual:
          type: integer
        sanMaximo:
          type: integer
        forca:
          type: integer
        agilidade:
          type: integer
        intelecto:
          type: integer
        vigor:
          type: integer
        presenca:
          type: integer
        defesa:
          type: integer
        historia:
          type: string
        pericias:
          type: array
          items:
            type: string
        rituals:
          type: array
          items:
            type: string
        inventario:
          type: array
          items:
            type: object
        classe:
          $ref: "#/components/schemas/RpgClass"
        origem:
          $ref: "#/components/schemas/Origin"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - nome
        - nivel
        - nex
    CreateCharacterBody:
      type: object
      properties:
        nome:
          type: string
        classeId:
          type: string
        origemId:
          type: string
        nivel:
          type: integer
          default: 1
        nex:
          type: integer
          default: 0
        patente:
          type: string
        forca:
          type: integer
          default: 1
        agilidade:
          type: integer
          default: 1
        intelecto:
          type: integer
          default: 1
        vigor:
          type: integer
          default: 1
        presenca:
          type: integer
          default: 1
        historia:
          type: string
        pericias:
          type: array
          items:
            type: string
      required:
        - nome
    UpdateCharacterBody:
      type: object
      properties:
        nome:
          type: string
        classeId:
          type: string
        origemId:
          type: string
        nivel:
          type: integer
        nex:
          type: integer
        patente:
          type: string
        pvAtual:
          type: integer
        peAtual:
          type: integer
        sanAtual:
          type: integer
        forca:
          type: integer
        agilidade:
          type: integer
        intelecto:
          type: integer
        vigor:
          type: integer
        presenca:
          type: integer
        defesa:
          type: integer
        historia:
          type: string
        pericias:
          type: array
          items:
            type: string
        rituals:
          type: array
          items:
            type: string
        inventario:
          type: array
          items:
            type: object
    RpgClass:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        descricao:
          type: string
        pvInicial:
          type: integer
        pvPorNivel:
          type: integer
        peInicial:
          type: integer
        pePorNivel:
          type: integer
        sanInicial:
          type: integer
        sanPorNivel:
          type: integer
        periciasTreindasBase:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - nome
...
    Origin:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        descricao:
          type: string
        periciasConcedidas:
          type: array
          items:
            type: string
        poderConcedido:
          type: string
        createdAt:
          type: string
      required:
        - id
        - nome
...
    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: string
      required:
        - code
        - code_verifier
        - redirect_uri
        - state
    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
...