# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CharacterInventarioItem } from "./characterInventarioItem";
import type { Origin } from "./origin";
import type { RpgClass } from "./rpgClass";

export interface Character {
  id: string;
  userId: string;
  nome: string;
  classeId?: string;
  origemId?: string;
  nivel: number;
  nex: number;
  patente?: string;
  pvAtual?: number;
  pvMaximo?: number;
  peAtual?: number;
  peMaximo?: number;
  sanAtual?: number;
  sanMaximo?: number;
  forca?: number;
  agilidade?: number;
  intelecto?: number;
  vigor?: number;
  presenca?: number;
  defesa?: number;
  historia?: string;
  pericias?: string[];
  rituals?: string[];
  inventario?: CharacterInventarioItem[];
  classe?: RpgClass;
  origem?: Origin;
  createdAt?: string;
  updatedAt?: string;
}