# lib/api-zod/src/generated/types/createCharacterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCharacterBody {
  nome: string;
  classeId?: string;
  origemId?: string;
  nivel?: number;
  nex?: number;
  patente?: string;
  forca?: number;
  agilidade?: number;
  intelecto?: number;
  vigor?: number;
  presenca?: number;
  historia?: string;
  pericias?: string[];
}