# lib/api-zod/src/generated/types/origin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface Origin {
  id: string;
  nome: string;
  descricao?: string;
  periciasConcedidas?: string[];
  poderConcedido?: string;
  createdAt?: string;
}