# lib/api-zod/src/generated/types/updateCharacterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ordo Realitas RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCharacterBodyInventarioItem } from "./updateCharacterBodyInventarioItem";

export interface UpdateCharacterBody {
  nome?: string;
  classeId?: string;
  origemId?: string;
  nivel?: number;
  nex?: number;
  patente?: string;
  pvAtual?: number;
  peAtual?: number;
  sanAtual?: number;
  forca?: number;
  agilidade?: number;
  intelecto?: number;
  vigor?: number;
  presenca?: number;
  defesa?: number;
  historia?: string;
  pericias?: string[];
  rituals?: string[];
  inventario?: UpdateCharacterBodyInventarioItem[];
}